Update teamsphisher.py with new Bypass Technique
There is a new Teams message warning bypass which requires users to send a message and remove themselves from the group chat. TeamsPhisher already invites the user twice, so the only modification was creating a new function. Users of TeamsPhisher may want to keep the message warning depending on the pretext and what they are attempting to accomplish. I have not created an option to do this yet.

Main Change:
Created a new function called removeExternalUser which deletes the "targetInfo" member once a chat is created. This function is called on line 804 once a new real thread/message is sent. 

removeExternalUser(skypeToken, senderInfo, threadID, targetInfo)

Bypass Details
https://badoption.eu/blog/2024/01/12/teams5.html